requirements.txt
python-binance
python-dotenv~=1.0.0